<REFACTOR> 프로젝트내 패키지 이름 convention을 유지하기 위해 exceptions를 exception으로 Rename
